refactor(gui-v2): make create and update project pages subpages of dashboard
