--- packages/nc-gui-v2/pages/project/index.vue
@@ -1,5 +1,0 @@
-<template>
-  <NuxtLayout>
-    <NuxtPage />
-  </NuxtLayout>
-</template>
